can reproduce NCEP route with classes split between IsoBased, RoutingAlgTime/FuelMin
